Skip unnecessary frequent attachment verification

+++ Planet/Entities/MyPlanetModel.swift
@@ -88,6 +88,7 @@ class MyPlanetModel: Equatable, Hashable, Identifiable, ObservableObject, Codabl
     @Published var articles: [MyArticleModel]! = nil
 
     var ops: [String: Date] = [:]
+    var attachmentsLastVerified: Date? = nil
 
     var tags: [String: String]? = [:]
     /// Array of remote sources for aggregating content
@@ -1573,46 +1574,53 @@ class MyPlanetModel: Equatable, Hashable, Identifiable, ObservableObject, Codabl
         }
         var publicArticles = articles.filter { $0.articleType == .blog }.map { $0.publicArticle }
         // Check if all public articles have complete attachments
-        debugPrint("About to check all attachments for planet \(name)")
-        for article in articles {
-            var shouldRemoveArticle = false
-            if let attachments = article.attachments {
-                debugPrint("Attachments check: \(article.id) should have \(attachments.count) attachments")
-                for attachment in attachments {
-                    var shouldRemove = false
-                    let attachmentPath = article.publicBasePath.appendingPathComponent(attachment)
-                    if !FileManager.default.fileExists(atPath: attachmentPath.path) {
-                        shouldRemoveArticle = true
-                        debugPrint("Attachments check: \(attachmentPath) not found")
-                    } else {
-                        let fileSize = try FileManager.default.attributesOfItem(atPath: attachmentPath.path)[.size] as? Int ?? 0
-                        if fileSize == 0 {
-                            debugPrint("Attachments check: \(attachmentPath) is empty")
-                            shouldRemove = true
+        if let verified = attachmentsLastVerified, verified > Date().addingTimeInterval(-60 * 10) {
+            debugPrint("Attachments check: skipping, last verified \(verified)")
+        }
+        else {
+            debugPrint("Attachments check: running")
+            attachmentsLastVerified = Date()
+            debugPrint("Attachments check: About to check all attachments for planet \(name)")
+            for article in articles {
+                var shouldRemoveArticle = false
+                if let attachments = article.attachments {
+                    debugPrint("Attachments check: \(article.id) should have \(attachments.count) attachments")
+                    for attachment in attachments {
+                        var shouldRemove = false
+                        let attachmentPath = article.publicBasePath.appendingPathComponent(attachment)
+                        if !FileManager.default.fileExists(atPath: attachmentPath.path) {
+                            shouldRemoveArticle = true
+                            debugPrint("Attachments check: \(attachmentPath) not found")
                         } else {
-                            if fileSize < 1000 {
-                                let fileDataString = try String(contentsOf: attachmentPath)
-                                if fileDataString.contains("no link named") && fileDataString.contains("under") {
-                                    debugPrint("Attachments check: \(attachmentPath) is a broken link")
-                                    shouldRemove = true
+                            let fileSize = try FileManager.default.attributesOfItem(atPath: attachmentPath.path)[.size] as? Int ?? 0
+                            if fileSize == 0 {
+                                debugPrint("Attachments check: \(attachmentPath) is empty")
+                                shouldRemove = true
+                            } else {
+                                if fileSize < 1000 {
+                                    let fileDataString = try String(contentsOf: attachmentPath)
+                                    if fileDataString.contains("no link named") && fileDataString.contains("under") {
+                                        debugPrint("Attachments check: \(attachmentPath) is a broken link")
+                                        shouldRemove = true
+                                    } else {
+                                        debugPrint("Attachments check: \(attachmentPath) is OK")
+                                    }
                                 } else {
                                     debugPrint("Attachments check: \(attachmentPath) is OK")
                                 }
-                            } else {
-                                debugPrint("Attachments check: \(attachmentPath) is OK")
                             }
                         }
-                    }
-                    if shouldRemove {
-                        shouldRemoveArticle = true
-                        debugPrint("Attachments check: \(attachmentPath) is broken, removing")
-                        try FileManager.default.removeItem(at: attachmentPath)
+                        if shouldRemove {
+                            shouldRemoveArticle = true
+                            debugPrint("Attachments check: \(attachmentPath) is broken, removing")
+                            try FileManager.default.removeItem(at: attachmentPath)
+                        }
                     }
                 }
-            }
-            if shouldRemoveArticle {
-                debugPrint("Attachments check: \(article.id) has broken attachments, removing article from public site")
-                publicArticles.removeAll { $0.id == article.id }
+                if shouldRemoveArticle {
+                    debugPrint("Attachments check: \(article.id) has broken attachments, removing article from public site")
+                    publicArticles.removeAll { $0.id == article.id }
+                }
             }
         }
         let publicPlanet = PublicPlanetModel(
